+++ test/dummy/app/models/user.rb
@@ -1,6 +1,17 @@
 class User < ApplicationRecord
-  # Include default devise modules. Others available are:
-  # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
   devise :database_authenticatable, :registerable,
          :recoverable, :rememberable, :validatable
+
+  recording_studio_recordable(
+    label: "User",
+    plural_label: "Users",
+    root: false,
+    allowed_parent_types: ["Workspace"]
+  )
+
+  validates :name, presence: true
+
+  def readonly?
+    false
+  end
 end